feat(atari): wave 3 — all 42 remaining ALE games (103 total, every supported ROM)
adventure air_raid atlantis2 backgammon basic_math blackjack carnival casino
crossbow donkey_kong earthworld entombed et flag_capture frogger galaxian
hangman haunted_house human_cannonball journey_escape kaboom keystone_kapers
king_kong koolaid laser_gates lost_luggage mario_bros miniature_golf mr_do
othello pacman sir_lancelot space_war superman tetris tic_tac_toe_3d trondead
turmoil video_checkers video_chess video_cube word_zapper.
Not portable: pitfall2 (DPC chip); combat/joust/maze_craze/warlords (no ALE
support exists).

New mechanisms:
- select_until gained a press-duration (ET needs ~25-frame SELECT presses)
  and ~20 entries: ALE's default mode = FIRST of getAvailableModes (not 0) —
  flag_capture boots to mode 8, galaxian/air_raid/video_checkers to 1.
- fire_until: FIRE-mash start for Mario Bros (title FIRE polling window is
  timing-sensitive; single presses miss nondeterministically; mash 2-on/28-
  off until the lives byte latches).
- game_aux widened to Int64 (MiniatureGolf packs five counters; Hangman
  four). _sb signed RAM read (Backgammon pieces, WordZapper rounds).

Traps fixed:
- Pacman ($CC) and HumanCannonball ($B6) mode bytes double as score bytes
  (read 0 post-RESET) — a select loop scrambles them into 2-player mode;
  both boot in the default mode, no select needed.
- Galaxian: ALE's treat-any-drop-as-1M-wrap false-fires on our game-over
  RAM transient; wrap now gated on prev_score > 990000.
- word_zapper's title screen read as terminal every step (round timer 0
  while wall clock ticks), starving the env of the in-game start press —
  game_aux latches 'timer was seen running'.

Known-broken (flagged in docs/ATARI_ALE_GAMES_IMPORT.md): haunted_house
(boot freeze in the display kernel's divide loop at $F44F with RAM $0F=0,
VSYNC stops), casino (input-disabled bit set at boot; paddle cart).

Validation: same-regime ale-py per game — 39/42 pass; entombed deterministic
script frame-match (total 3==3, first reward step 6 both), backgammon /
basic_math / flag_capture / human_cannonball exact; donkey_kong frames near
pixel-identical to ALE (its smoke 0 is rare-event variance, ALE scored once
in 2000 steps). Waves 1-2 rows bit-identical; tests green.

+++ mojo_rl/envs/atari/atari_state.mojo
@@ -104,8 +104,9 @@ struct AtariState(Copyable, Movable):
     var frame_number: UInt32  # Total frames elapsed
     # Per-game persistent scratch for game_signals (ALE settings keep private
     # state across steps: ChopperCommand's started latch, DarkChambers' last
-    # level). Reset to 0 with the rest of the state on env.reset().
-    var game_aux: Int32
+    # level, MiniatureGolf's five packed counters). Reset to 0 with the rest
+    # of the state on env.reset().
+    var game_aux: Int64
 
     # ========================================================================
     # RAM (128 bytes)
